Route remote plugin config toggles outside ConfigApi
